+++ src/api/routes/v1/user.route.ts
@@ -183,6 +183,7 @@ router
   .delete(authorize(LOGGED_USER), controller.remove);
 
 router.route('/:userId/notes').get(authorize(LOGGED_USER), controller.listUserNotes);
+router.route('/:userId/notes').post(authorize(LOGGED_USER), controller.createNote);
 router.route('/:userId/notes/:noteId').delete(authorize(LOGGED_USER), controller.deleteUserNote);
 
 module.exports = router;